+++ molt-book/src/changes.md
@@ -5,6 +5,7 @@
 * Added the [**molt_throw!**](embed/molt_result.md#result-macros) macro.
 * Improved the API documentation for `molt_ok!` and `molt_err!`.
 * Added `Exception::error_code` and `Exception::error_info`, to streamline using exceptions.
+* Added the [**string**](ref/string.md) command
 
 ## New in Molt 0.3.0